Streamlined code, added warning note
Previous fix wasn't actually fixing aything, since if lastProperty was anywhere in the properties chain
it would NOT be propagated correctly. I could preserve lastProperty only if it was referencing the topmost
properties (i.e. a variable). Thus, the check was pretty much useless
I've opted to remove the explicit check and retained the actual fix (not concatenating with empty Property objects)